Wall Sections
Now, I have added a wall section for the Start, About and Levels, where only the About and Levels have a go back arrow (on which we shall continue working on).
Pressing them will trigger a little animation of a wall expanding from the left side of the screen.
The next part is working on the actual game...as for the About section and Level section depend on this.