Main Screen
I have made some changes to the main screen
For starters, I have removed the sky background...then I added the more appropriate raining background as I believe it sets the mood for the character/story...
Then the buttons were redesigned. They have now a hover effect. They are now parts of a scroll ( 📜 ).
For now, it looks ok...
In the future, we must design the logo in a mossy swamp-infested rock.
But for now, it serves our interests...