Chests 2
I have added a random item generator.
For its limited to 4 gems, but it can get quite nice as I build up.
And I have generated the inventory visually.
The items in the chests are generated and positioned randomly.
------------
Now, we need to bind keys to pick up the items in our inventory.
The key bind should be 'T'.
And we also need to do it the other way around, in which case the key bind should be 'R'.
---------